requirements.txt
appdirs==1.4.3
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
cssselect==1.1.0
decorator==4.4.2
fake-useragent==0.1.11
future==0.18.2
html5lib==1.0.1
idna==2.9
lxml==4.5.0
parse==1.15.0
pycparser==2.20
pyee==7.0.1
pyppeteer==0.0.25
pyquery==1.4.1
pyTelegramBotAPI==3.6.7
python-telegram-bot==12.6.1
requests==2.23.0
requests-html==0.10.0
six==1.14.0
soupsieve==2.0
tornado==6.0.4
tqdm==4.45.0
urllib3==1.25.9
w3lib==1.21.0
webencodings==0.5.1
websockets==8.1
